fix(product): split the oversized launch and pending-sidebar test files (PRO-230)
The concurrency cases this slice added pushed two suites past the repo
file-size gate (use-home-next-launch.test.tsx 790, max 600;
pending-sidebar-projection.test.ts 664). Split each along the seam that was
already there — single-launch behaviour stays in the original file, the
concurrent-launch behaviour moves to a colocated sibling — and pull the shared
render wrapper into `use-home-next-launch.test-support.tsx`. Module mocks and
the store reset stay per file, because vitest registers mocks per file and only
a test file may write store state directly. No case was dropped or weakened:
the same 14 + 15 assertions run, just in two files each.
