+++ apps/packages/product-client/src/hooks/home/workflows/use-home-next-launch.test.tsx
@@ -39,6 +39,11 @@ import type {
 } from "#product/lib/domain/home/home-next-launch";
 import { CoworkThreadLaunchProvider } from "#product/providers/CoworkThreadLaunchProvider";
 
+import {
+  renderHomeNextLaunch,
+  sessionIdForAttempt,
+} from "#product/hooks/home/workflows/use-home-next-launch.test-support";
+
 const mocks = vi.hoisted(() => {
   const createThreadFromSelection = vi.fn();
   return {
@@ -121,13 +126,10 @@ vi.mock("#product/hooks/sessions/workflows/use-session-interaction-resolution-ac
   }),
 }));
 
-function launchWrapper({ children }: { children: ReactNode }) {
-  return <CoworkThreadLaunchProvider>{children}</CoworkThreadLaunchProvider>;
-}
-
-function renderHomeNextLaunch() {
-  return renderHook(() => useHomeNextLaunch(), { wrapper: launchWrapper });
-}
+// Routing half: which workflow a Home launch target reaches, and how the
+// launch intent is scoped to its own pending attempt. The concurrency cases
+// (two live launches, the cap, failure announcement, prompt dedupe) live in
+// `use-home-next-launch.concurrency.test.tsx`; the file-size gate splits them.
 
 describe("useHomeNextLaunch", () => {
   beforeEach(() => {
@@ -445,346 +447,4 @@ describe("useHomeNextLaunch", () => {
     expect(mocks.createLocalWorkspaceAndEnterWithResult).not.toHaveBeenCalled();
     expect(mocks.createWorktreeAndEnterWithResult).not.toHaveBeenCalled();
   });
-  const worktreeTarget = {
-    kind: "worktree" as const,
-    repoRootId: "repo-root-1",
-    sourceWorkspaceId: null,
-    baseBranch: "main",
-    defaultBranch: "main",
-  };
-
-  function pendingWorktreeEntry(attemptId: string) {
-    return buildSubmittingPendingWorkspaceEntry({
-      attemptId,
-      selectedWorkspaceId: null,
-      source: "worktree-created",
-      displayName: attemptId,
-      repoLabel: "repo",
-      baseBranchName: "main",
-      request: {
-        kind: "worktree",
-        input: {
-          repoRootId: "repo-root-1",
-          sourceWorkspaceId: null,
-          baseBranch: "main",
-          defaultBranch: "main",
-        },
-      },
-    });
-  }
-
-  function cloudPendingEntry(attemptId: string) {
-    return buildSubmittingPendingWorkspaceEntry({
-      attemptId,
-      selectedWorkspaceId: null,
-      source: "cloud-created",
-      displayName: attemptId,
-      repoLabel: "proliferate-ai/proliferate",
-      baseBranchName: "main",
-      request: {
-        kind: "cloud",
-        input: {
-          gitOwner: "proliferate-ai",
-          gitRepoName: "proliferate",
-          baseBranch: "main",
-          branchName: attemptId,
-          generatedName: true,
-        },
-      },
-    });
-  }
-
-  function registerPendingEntry(entry: ReturnType<typeof buildSubmittingPendingWorkspaceEntry>) {
-    useSessionSelectionStore.setState((state) => ({
-      pendingWorkspaces: upsertPendingWorkspaceEntry(state.pendingWorkspaces, entry),
-    }));
-  }
-
-  it("runs two Home launches at once, each prompt routed to its own attempt (PRO-230)", async () => {
-    const finishByAttemptId = new Map<string, () => void>();
-    const entriesByAttemptId = new Map<
-      string,
-      ReturnType<typeof buildSubmittingPendingWorkspaceEntry>
-    >();
-    const sessionIdFor = (attemptId: string) => `client-session:codex:${attemptId}`;
-
-    mocks.createWorktreeAndEnterWithResult.mockImplementation((_input, options) => {
-      const attemptId: string = options.attemptId;
-      const sessionId = sessionIdFor(attemptId);
-      const entry = pendingWorktreeEntry(attemptId);
-      entriesByAttemptId.set(attemptId, entry);
-      putSessionRecord(createEmptySessionRecord(sessionId, "codex", {
-        workspaceId: buildPendingWorkspaceUiKey(entry),
-        materializedSessionId: null,
-        modelId: "gpt-5.4",
-      }));
-      // Entering the shell moves selection onto the newest attempt, exactly as
-      // a second submit does in production. The first launch must keep routing
-      // to its own attempt anyway.
-      useSessionSelectionStore.getState().enterPendingWorkspaceShell(entry, {
-        initialActiveSessionId: sessionId,
-      });
-      return new Promise((resolve) => {
-        finishByAttemptId.set(attemptId, () => resolve({
-          workspaceId: `workspace-${attemptId}`,
-          projectedSessionId: sessionId,
-        }));
-      });
-    });
-
-    const { result } = renderHomeNextLaunch();
-
-    let firstLaunch: Promise<HomeNextLaunchOutcome> = Promise.resolve("refused");
-    let secondLaunch: Promise<HomeNextLaunchOutcome> = Promise.resolve("refused");
-    await act(async () => {
-      firstLaunch = result.current.launch({
-        text: "first prompt",
-        modelSelection: { kind: "codex", modelId: "gpt-5.4" },
-        modeId: null,
-        launchControlValues: {},
-        target: worktreeTarget,
-      });
-      await Promise.resolve();
-      secondLaunch = result.current.launch({
-        text: "second prompt",
-        modelSelection: { kind: "codex", modelId: "gpt-5.4" },
-        modeId: null,
-        launchControlValues: {},
-        target: worktreeTarget,
-      });
-      await Promise.resolve();
-    });
-
-    // Both launches are live at the same time: neither the store nor the
-    // pending registry collapsed them into one.
-    const liveIntents = launchIntents(useChatLaunchIntentStore.getState());
-    expect(liveIntents).toHaveLength(2);
-    expect(new Set(liveIntents.map((intent) => intent.text)))
-      .toEqual(new Set(["first prompt", "second prompt"]));
-    const attemptIds = [...finishByAttemptId.keys()];
-    expect(attemptIds).toHaveLength(2);
-    expect(
-      useSessionSelectionStore.getState().pendingWorkspaces.attemptOrder,
-    ).toEqual(attemptIds);
-
-    let results: HomeNextLaunchOutcome[] = [];
-    await act(async () => {
-      for (const finish of finishByAttemptId.values()) {
-        finish();
-      }
-      results = await Promise.all([firstLaunch, secondLaunch]);
-    });
-
-    expect(results).toEqual(["launched", "launched"]);
-    const [firstAttemptId, secondAttemptId] = attemptIds as [string, string];
-    expect(
-      getPromptOutboxEntriesForSession(sessionIdFor(firstAttemptId))
-        .map((entry) => entry.text),
-    ).toEqual(["first prompt"]);
-    expect(
-      getPromptOutboxEntriesForSession(sessionIdFor(secondAttemptId))
-        .map((entry) => entry.text),
-    ).toEqual(["second prompt"]);
-    expect(useChatLaunchIntentStore.getState().intentOrder).toEqual([]);
-  });
-
-  it("refuses a launch past the concurrent cap without minting an intent", async () => {
-    let registry = EMPTY_PENDING_WORKSPACE_REGISTRY;
-    for (let index = 0; index < MAX_CONCURRENT_PENDING_LAUNCHES; index += 1) {
-      registry = upsertPendingWorkspaceEntry(registry, pendingWorktreeEntry(`attempt-${index}`));
-    }
-    useSessionSelectionStore.setState({ pendingWorkspaces: registry });
-
-    const { result } = renderHomeNextLaunch();
-    let outcome: HomeNextLaunchOutcome = "launched";
-    await act(async () => {
-      outcome = await result.current.launch({
-        text: "one too many",
-        modelSelection: { kind: "codex", modelId: "gpt-5.4" },
-        modeId: null,
-        launchControlValues: {},
-        target: worktreeTarget,
-      });
-    });
-
-    expect(outcome).toBe("refused");
-    expect(mocks.showToast).toHaveBeenCalledWith(
-      "Too many workspaces starting. Wait for one to finish.",
-      "info",
-    );
-    expect(mocks.createWorktreeAndEnterWithResult).not.toHaveBeenCalled();
-    expect(useChatLaunchIntentStore.getState().intentOrder).toEqual([]);
-    expect(
-      useSessionSelectionStore.getState().pendingWorkspaces.attemptOrder,
-    ).toHaveLength(MAX_CONCURRENT_PENDING_LAUNCHES);
-  });
-
-  // PR #1870 review finding 1 (blocking): the cloud branch was the one launch
-  // path that never learned its own attempt id, so its failure handling
-  // resolved "the attended attempt" — under concurrency, another launch's. A
-  // cloud failure then bound this launch's intent to the attended launch's
-  // attempt and workspace, which is how one launch's failure pane surfaced over
-  // another launch's shell.
-  it("scopes a failed cloud launch to its own attempt, not the attended one (PRO-230)", async () => {
-    // The attended launch is another cloud launch, which is what makes the
-    // fallback dangerous: any cloud-created entry matches a cloud target, so
-    // "the attended attempt" resolves to it.
-    const attended = cloudPendingEntry("attended-other-launch");
-    const attendedWithWorkspace = { ...attended, workspaceId: "workspace-attended" };
-    let cloudAttemptId: string | null = null;
-
-    mocks.createCloudWorkspaceAndEnterWithResult.mockImplementation(
-      async (_target: unknown, options: { attemptId: string }) => {
-        cloudAttemptId = options.attemptId;
-        // The cloud attempt registers itself, exactly as beginPendingWorkspace
-        // does inside the real create flow's synchronous prefix.
-        registerPendingEntry(cloudPendingEntry(options.attemptId));
-        // Meanwhile the user is attending a different launch entirely.
-        registerPendingEntry(attendedWithWorkspace);
-        useSessionSelectionStore.getState().enterPendingWorkspaceShell(attendedWithWorkspace, {
-          initialActiveSessionId: null,
-        });
-        return { status: "interrupted", failureMessage: "Billing gate" };
-      },
-    );
-
-    const { result } = renderHomeNextLaunch();
-    let outcome: HomeNextLaunchOutcome = "launched";
-    await act(async () => {
-      outcome = await result.current.launch({
-        text: "launch in cloud",
-        modelSelection: { kind: "codex", modelId: "gpt-5.4" },
-        modeId: null,
-        launchControlValues: {},
-        target: {
-          kind: "cloud",
-          gitOwner: "proliferate-ai",
-          gitRepoName: "proliferate",
-          baseBranch: "main",
-        },
-      });
-    });
-
-    expect(outcome).toBe("not-started");
-    expect(cloudAttemptId).not.toBeNull();
-    const failedIntent = launchIntents(useChatLaunchIntentStore.getState())[0];
-    expect(failedIntent?.attemptId).toBe(cloudAttemptId);
-    expect(failedIntent?.attemptId).not.toBe(attended.attemptId);
-    // Nor may it inherit the attended launch's workspace, which is what made
-    // Back and dismiss act on the wrong launch.
-    expect(failedIntent?.materializedWorkspaceId).toBeNull();
-    expect(failedIntent?.failure?.retryMode).toBe("safe");
-  });
-
-  // PR #1870 review finding 5: the per-attempt notice and the launch-level
-  // toast both fired for one unattended failure, the second one telling the
-  // user their prompt was back in a composer they were not looking at.
-  it("announces an unattended failure once, and an attended one in the shell", async () => {
-    const failWhileUnattended = async (_input: unknown, options: { attemptId: string }) => {
-      registerPendingEntry({
-        ...pendingWorktreeEntry(options.attemptId),
-        stage: "failed",
-        errorMessage: "Failed to create worktree.",
-      });
-      throw new Error("Failed to create worktree.");
-    };
-    mocks.createWorktreeAndEnterWithResult.mockImplementation(failWhileUnattended);
-
-    const { result } = renderHomeNextLaunch();
-    await act(async () => {
-      await result.current.launch({
-        text: "fail out of sight",
-        modelSelection: { kind: "codex", modelId: "gpt-5.4" },
-        modeId: null,
-        launchControlValues: {},
-        target: worktreeTarget,
-      });
-    });
-
-    // The attempt's own notice owns this one, so the launch adds nothing.
-    expect(mocks.showErrorToast).not.toHaveBeenCalled();
-
-    mocks.createWorktreeAndEnterWithResult.mockImplementation(
-      async (_input: unknown, options: { attemptId: string }) => {
-        const failed = {
-          ...pendingWorktreeEntry(options.attemptId),
-          stage: "failed" as const,
-          errorMessage: "Failed to create worktree.",
-        };
-        registerPendingEntry(failed);
-        useSessionSelectionStore.getState().enterPendingWorkspaceShell(failed, {
-          initialActiveSessionId: null,
-        });
-        throw new Error("Failed to create worktree.");
-      },
-    );
-
-    await act(async () => {
-      await result.current.launch({
-        text: "fail while watching",
-        modelSelection: { kind: "codex", modelId: "gpt-5.4" },
-        modeId: null,
-        launchControlValues: {},
-        target: worktreeTarget,
-      });
-    });
-
-    expect(mocks.showErrorToast).toHaveBeenCalledTimes(1);
-    expect(mocks.showErrorToast.mock.calls[0]?.[0].headline).toBe("Work not started");
-  });
-
-  it("collapses the same prompt submitted twice but starts two different ones", async () => {
-    mocks.createWorktreeAndEnterWithResult.mockImplementation((_input, options) => {
-      const attemptId: string = options.attemptId;
-      const sessionId = sessionIdForAttempt(attemptId);
-      const entry = pendingWorktreeEntry(attemptId);
-      putSessionRecord(createEmptySessionRecord(sessionId, "codex", {
-        workspaceId: buildPendingWorkspaceUiKey(entry),
-        materializedSessionId: null,
-        modelId: "gpt-5.4",
-      }));
-      useSessionSelectionStore.getState().enterPendingWorkspaceShell(entry, {
-        initialActiveSessionId: sessionId,
-      });
-      return Promise.resolve({ workspaceId: `workspace-${attemptId}`, projectedSessionId: sessionId });
-    });
-
-    const { result } = renderHomeNextLaunch();
-    const submit = (text: string, target: HomeLaunchTarget = worktreeTarget) =>
-      result.current.launch({
-        text,
-        modelSelection: { kind: "codex" as const, modelId: "gpt-5.4" },
-        modeId: null,
-        launchControlValues: {},
-        target,
-      });
-
-    let repeated: HomeNextLaunchOutcome = "launched";
-    await act(async () => {
-      await submit("ship it");
-      repeated = await submit("ship it");
-    });
-    expect(mocks.createWorktreeAndEnterWithResult).toHaveBeenCalledTimes(1);
-    // The suppressed submit is not a failure: the launch it collapsed into is
-    // running, so the composer must not hand the prompt back (review finding 7).
-    expect(repeated).toBe("duplicate");
-
-    await act(async () => {
-      await submit("ship something else");
-    });
-    expect(mocks.createWorktreeAndEnterWithResult).toHaveBeenCalledTimes(2);
-
-    // Same prompt, different repository, inside the same second: two launches.
-    await act(async () => {
-      await submit("ship something else", {
-        ...worktreeTarget,
-        repoRootId: "repo-root-2",
-      });
-    });
-    expect(mocks.createWorktreeAndEnterWithResult).toHaveBeenCalledTimes(3);
-  });
 });
-
-function sessionIdForAttempt(attemptId: string) {
-  return `client-session:codex:${attemptId}`;
-}
